requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boto3==1.29.2
botocore==1.32.2
certifi==2023.7.22
charset-normalizer==3.3.2
frozenlist==1.4.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
twilio==8.10.2
urllib3==1.26.18
yarl==1.9.2
